Fix license headers: Should have been AGPLv3+, not GPLv2+
I'm not sure why one files (written by Philipp)
did contain a GPLv2+ header, instead of the AGPLv3+ which is the
actual overall project license.  I consider it a mistake.

In any case, any copyrightable contribution to those files was done by
sysmocom employees, so I as managing directory can legally make a
license change, whther or not it was a mistake early on or not.

Change-Id: Ie28c32022a9cbc42b55195f7dbb513228857a3cd