ranap_common_ran: remove unused ctx args
This commit is basically 8b6dad3331eab339a7ffea7c22ed1830fe9ced60
applied to ranap_common_ran instead of ranap_common_cn.

ranap_ran_rx_co has a void *ctx argument, that gets passed to:
* a callback function and
* to various decode functions in the same file.

As it is named "ctx", it looks like a talloc context. But the decode
functions don't use ctx at all and so in reality it is private userdata
for the callback. It is used as such by test/hnb-test-rua and in
osmo-msc.

Change-Id: I57b320c2fdd72d11b87507d0cd9c2e7b479e7fe4